EDIT: Fix formating left, fix error for value scale down divide correct

+++ CollAction/Services/Image/AmazonS3ImageService.cs
@@ -131,14 +131,15 @@ namespace CollAction.Services.Image
 
         private double GetScaleRatioForImage(Image<Rgba32> image)
         {
-            if (image.Width > _imageResizeThreshold || image.Height > _imageResizeThreshold){
+            if (image.Width > _imageResizeThreshold || image.Height > _imageResizeThreshold)
+            {
                 if (image.Width > image.Height) 
                 {
-                    return (_imageResizeThreshold/image.Width);
+                    return ( (double)_imageResizeThreshold/(double)image.Width);
                 }
                 else
                 {
-                    return (_imageResizeThreshold/image.Height);
+                    return ( (double)_imageResizeThreshold/(double)image.Height);
                 }
             }
